requirements.txt
# ── MCP Protocol ─────────────────────────────────────────────────────────────
mcp[cli]>=1.5.0

# ── PostgreSQL ────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9

# ── AI / LLM ─────────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Storage / ORM ─────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30

# ── Config Management ─────────────────────────────────────────────────────────
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# ── Background Scheduler (autonomous capture daemon) ──────────────────────────
apscheduler>=3.10.4

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
